Island: Move MongoMasqueradeRepository to agent_binary_service
