+++ monkey/monkey_island/cc/services/agent_binary_service/i_masquerade_repository.py
